refactor(console): extract shared schema field renderer
